(meta_lab): Add DumSegment

+++ cpp_impl/Schema.h
@@ -5,8 +5,8 @@
 #include <set>
 
 using ID_TYPE = int64_t;
-using NUM_TYPE = int8_t;
-using FTYPE_TYPE = int16_t;
+using NUM_TYPE = int64_t;
+using FTYPE_TYPE = int64_t;
 using TS_TYPE = int64_t;
 using MappingT = std::set<ID_TYPE>;
 /* using MappingT = std::vector<ID_TYPE>; */
